+++ hansen-cli/packages/template-react/CHANGELOG.md
@@ -1,5 +1,10 @@
 <!--
 
+## 1.3.0
+### Minor Changes
+
+- eslint
+
 ## 1.2.0
 ### Minor Changes