+++ hansen-cli/packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hansen-cli/utils
 
+## 1.2.0
+
+### Minor Changes
+
+- eslint
+
 ## 1.1.2
 
 ### Patch Changes